% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.d.diag.R
\name{ssa.d.diag}
\alias{ssa.d.diag}
\title{Direct method (D) for nu-diagonalized systems}
\usage{
ssa.d.diag(a, nu)
}
\arguments{
\item{a}{vector of evaluated propensity functions.}

\item{nu}{state-change matrix.}
}
\value{
A list with two elements:
\itemize{
\item the time leap (\code{tau}) and
\item the realized state change vector (\code{nu_j}).
}
}
\description{
Direct method (D) for nu-diagonalized systems.
}
\details{
Performs one time step using the Direct method. It is usually called
from within \code{\link[=ssa]{ssa()}}, but can be invoked directly, see \code{\link[=ssa.d]{ssa.d()}} for Examples.
}
\seealso{
\code{\link[=ssa.d]{ssa.d()}}
}
\keyword{datagen}
\keyword{misc}
\keyword{ts}
