% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.d.R
\name{ssa.d}
\alias{ssa.d}
\title{Direct method (D)}
\usage{
ssa.d()
}
\description{
Direct method implementation of the \acronym{SSA} as described by Gillespie (1977).
Should be passed as \code{method} argument for \code{ssa()}.
}
\examples{
ssa.d()

}
\references{
Gillespie (1977)
}
\seealso{
\link{GillespieSSA-package}, \code{\link[=ssa]{ssa()}}
}
\keyword{datagen}
\keyword{misc}
\keyword{ts}
