% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.nutiling.R
\name{ssa.nutiling}
\alias{ssa.nutiling}
\title{Direct method nu-diagonalization mapping}
\usage{
ssa.nutiling(a, nu, j)
}
\arguments{
\item{a}{vector of evaluated propensity functions.}

\item{nu}{state-change matrix.}

\item{j}{Reaction index to map}
}
\value{
The virtual realized state change vector (\code{nu_j}).
}
\description{
Auxiliary function for \code{ssa.d.diag} performing virtual mapping of nu-diagonalized systems.
}
\seealso{
\code{\link[=ssa.d.diag]{ssa.d.diag()}}, \code{\link[=ssa.d]{ssa.d()}}
}
\keyword{datagen}
\keyword{misc}
\keyword{ts}
