% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.etl.R
\name{ssa.etl}
\alias{ssa.etl}
\title{Explicit tau-leap method (ETL)}
\usage{
ssa.etl(tau = 0.3)
}
\arguments{
\item{tau}{tau-leap.}
}
\description{
Explicit tau-leap method implementation of the \acronym{SSA} as described by
Gillespie (2001). Should be passed as \code{method} argument for \code{ssa()}.
}
\details{
Performs one time step using the Explicit tau-leap method.
Intended to be invoked by \code{\link[=ssa]{ssa()}}.
}
\examples{
ssa.etl(tau = .1)

}
\references{
Gillespie (2001)
}
\seealso{
\link{GillespieSSA-package}, \code{\link[=ssa]{ssa()}}
}
\keyword{datagen}
\keyword{misc}
\keyword{ts}
